requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.11
fonttools==4.56.0
h11==0.14.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.1
matplotlip==0.2
numba==0.61.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.46.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
starlette==0.46.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
xgboost-cpu==2.1.4
